Add files via upload
After comparing several classification models the best ICO project success prediction was obtained with the Random Forest algorithm. This classification model gave an efficient performance lying closer to the ideal classifier. This is mainly due to the versatile nature of the algorithm wherein it reduces overfitting and mechanizes the missing values in the noisy data.